pyproject.toml
[project]
name = "chaostooling-oss"
version = "0.1.0"
description = "Monorepo for Chaos Toolkit extensions and tooling"
authors = [{name = "Morgan Wigge", email = "morgan@wigge.nu"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}

dependencies = [
    # Core
    "chaostoolkit-lib>=1.42.1",
    
    # DB Extension
    "psycopg2-binary>=2.9.0",
    "mysql-connector-python>=8.0.0",
    "pyodbc>=4.0.39",
    "pymongo>=4.0.0",
    "redis>=4.5.0",
    "cassandra-driver>=3.28.0",
    "kafka-python>=2.0.2",
    "pika>=1.3.0",
    "stomp.py>=8.1.0",
    "requests>=2.31.0",
    "psutil>=5.9.0",

    # Compute Extension
    # (psutil included above)

    # Reporting Extension
    "jinja2>=3.0.0",
    "weasyprint>=60.0",
    "pandas>=2.0.0",

    # OTEL Extension
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-semantic-conventions>=0.45b0",
    "opentelemetry-exporter-otlp>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=0.45b0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "opentelemetry-instrumentation>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-requests>=0.45b0",
    "opentelemetry-instrumentation-urllib3>=0.45b0",
    "opentelemetry-propagator-b3>=1.24.0",
    "opentelemetry-propagator-jaeger>=1.24.0",
    "opentelemetry-exporter-gcp-trace>=1.6.0",
    "opentelemetry-resourcedetector-gcp>=1.6.0a0",
    "opentelemetry-propagator-gcp>=1.6.0",
    "opentelemetry-exporter-prometheus>=0.45b0",
    "opentelemetry-sdk-extension-aws>=2.0.1",
    "opentelemetry-propagator-aws-xray>=1.0.1",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-cov>=5.0.0",
    "mypy>=1.19.0",
    "black>=25.11.0",
    "ruff>=0.14.7",
    "isort",
    "bandit>=1.7.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
namespaces = false
where = [
    "chaostooling-extension-db",
    "chaostooling-otel",
    "chaostooling-extension-compute",
    "chaostooling-extension-network",
    "chaostooling-extension-app",
    "chaostooling-reporting"
]

# ============================================================================
# PYTEST CONFIGURATION
# ============================================================================

[tool.pytest.ini_options]
testpaths = ["chaostooling-generic/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]

# Markers for test categorization
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require database/external services)",
    "e2e: End-to-end tests (full workflow from discover to audit)",
    "slow: Tests that take >1 second to run",
    "db: Tests that require database connectivity",
    "flaky: Tests known to be flaky and need investigation",
]

# Coverage settings
[tool.coverage.run]
source = ["chaostooling-generic/chaosgeneric"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 95
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# ============================================================================
# CODE QUALITY TOOLS
# ============================================================================

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "N999",   # Invalid module name (packages use hyphens, not underscores) 
    "N803",   # Argument name should be lowercase (MongoDB API requires authSource camelCase)
    "N806",   # Variable in function should be lowercase (MongoDB API requires authSource camelCase)
    "E402",   # Module level import not at top (some scripts require dynamic imports)
]

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["chaosgeneric"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
show_error_context = true
pretty = true
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
]

# ============================================================================
# BANDIT SECURITY CONFIGURATION
# ============================================================================

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
tests = [
    "B201",
    "B301",
    "B302",
    "B303",
    "B304",
    "B305",
    "B306",
    "B307",
    "B308",
    "B309",
    "B310",
    "B311",
    "B312",
    "B313",
    "B314",
    "B315",
    "B316",
    "B317",
    "B318",
    "B319",
    "B320",
    "B321",
    "B322",
    "B323",
    "B324",
    "B325",
]
skips = []

